# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Web-Karir (Career Hub) Secure REST API",
        "description": "Dokumentasi REST API lengkap dan terproteksi untuk platform Web-Karir (Career Hub). Menggunakan Laravel Sanctum Bearer Token Authentication, Rate Limiting, dan Spatie Role-Based Access Control.",
        "contact": {
            "name": "Web-Karir Support Team",
            "email": "admin@webkarir.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "Web-Karir Primary API Server"
        }
    ],
    "paths": {
...
        "/notifications": {
            "get": {
                "tags": [
                    "Notifications & Salary"
                ],
                "summary": "Daftar Notifikasi User",
                "description": "Mengambil daftar notifikasi in-app bell pengguna.",
                "operationId": "7eca2b9e84a26784eb55b3fc6b697567",
                "responses": {
                    "200": {
                        "description": "Daftar notifikasi",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "message": "Notifikasi berhasil diambil.",
                                    "data": {
                                        "unread_count": 1,
                                        "notifications": [
                                            {
                                                "id": 1,
                                                "user_id": 1,
                                                "title": "Undangan Interview",
                                                "message": "Selamat! Anda diundang untuk mengikuti wawancara.",
                                                "is_read": false,
                                                "created_at": "2026-08-21T22:00:00.000000Z"
                                            }
                                        ]
                                    },
                                    "errors": null
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/notifications/{id}/read": {
            "post": {
                "tags": [
                    "Notifications & Salary"
                ],
                "summary": "Tandai Notifikasi Dibaca",
                "description": "Tandai notifikasi tertentu sebagai telah dibaca.",
                "operationId": "dbadc4694f2557775ca3c1996daba876",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notifikasi ditandai dibaca",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "message": "Notifikasi ditandai sudah dibaca.",
                                    "data": null,
                                    "errors": null
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/notifications/read-all": {
            "post": {
                "tags": [
                    "Notifications & Salary"
                ],
                "summary": "Tandai Semua Notifikasi Dibaca",
                "description": "Menandai seluruh notifikasi user sebagai telah dibaca.",
                "operationId": "925deb27b88c45df00bdf714699c3a68",
                "responses": {
                    "200": {
                        "description": "Seluruh notifikasi dibaca",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "message": "Semua notifikasi ditandai sudah dibaca.",
                                    "data": null,
                                    "errors": null
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/umk-lookup": {
            "get": {
                "tags": [
                    "Regional UMK & Benchmark"
                ],
                "summary": "Pencarian UMK 2026 Wilayah",
                "description": "Mencari besaran Nilai UMK Resmi 2026 berdasarkan lokasi kota/kabupaten.",
                "operationId": "940b056932eef816f95c339a32a06ba3",
                "parameters": [
                    {
                        "name": "location",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Jakarta Pusat"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data UMK ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "found": true,
                                    "city_district": "Kota Administrasi Jakarta Pusat",
                                    "province": "DKI JAKARTA",
                                    "umk_amount": 5729876,
                                    "formatted_umk": "Rp 5.729.876",
                                    "legal_decree": "KEPUTUSAN MENTERI KETENAGAKERJAAN REPUBLIK INDONESIA NOMOR 65 TAHUN 2026"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/regions/cities": {
            "get": {
                "tags": [
                    "Regional UMK & Benchmark"
                ],
                "summary": "Daftar Wilayah Kota/Kabupaten Autocomplete",
                "description": "Mengambil daftar kota/kabupaten di Indonesia untuk pilihan autocomplete.",
                "operationId": "3dba2910e52918bd274c4139f286b0b2",
                "parameters": [
                    {
                        "name": "query",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Bandung"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar kota/kabupaten",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": [
                                    {
                                        "city_district": "Kota Bandung",
                                        "province": "JAWA BARAT",
                                        "umk_amount": 4737678
                                    },
                                    {
                                        "city_district": "Kabupaten Bandung",
                                        "province": "JAWA BARAT",
                                        "umk_amount": 3972202
                                    }
                                ]
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your Sanctum Bearer Token to access protected API endpoints.",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}